% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multRegression.R
\name{adaLassoSelection}
\alias{adaLassoSelection}
\title{Run the selection method employing the adaptive lasso penalty (Zou, 2006)}
\usage{
adaLassoSelection(Xaug, Y, n, Omega0 = NULL)
}
\arguments{
\item{Xaug}{Design matrix of original data column augmented by knockoffs}

\item{Y}{vectorized response vector}

\item{n}{the number of samples}

\item{Omega0}{the estimated omega matrix}
}
\value{
The result of adaptive lasso method
}
\description{
Run the selection method employing the adaptive lasso penalty (Zou, 2006)
}
