% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multRegression.R
\name{LassoSelection}
\alias{LassoSelection}
\title{Run the selection method through Lasso penalty (L1)}
\usage{
LassoSelection(Xaug, Y, n, Omega0 = NULL)
}
\arguments{
\item{Xaug}{Design matrix of original data column augmented by knockoffs}

\item{Y}{vectorized response vector}

\item{n}{the number of samples}

\item{Omega0}{the estimated omega matrix}
}
\value{
The result of the lasso fit
}
\description{
Run the selection method through Lasso penalty (L1)
}
