% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCKF.R
\name{MRCKF}
\alias{MRCKF}
\title{Multiple Response Knockoff Filter}
\usage{
MRCKF(W, Ymat, q, penalty = "lasso")
}
\arguments{
\item{W}{OTU count matrix of dimension (n x p^*)}

\item{q}{target FDP (0,1)}

\item{penalty}{type of penalty for selection: "lasso", "alasso", or "scad"}

\item{Y}{matrix of responses (n x K)}
}
\value{
A list containing the following components:
\item{S}{The estimated selection matrix for B under the knockoff threshold}
\item{Sp}{The estimated selection matrix for B under the knockoff+ threshold}
}
\description{
Run the MRCKF procedure
}
\details{
The MRCKF method relies on the model-X formulation which assumes that the underlying
alr transformed data \eqn{X} follows a multivariate normal distribution. The model-X formulation generates
knockoff copies by estimating the underlying mean and covariance structure. Note that the MRCKF
method requires the input of the raw, non-compositional data \eqn{W}.

If the number of responses, \eqn{K=1}, then the MRCKF method defaults to a variation of the
compostional knockoff filter (CKF) by Srinivasan et al., 2020.

The knockoff importance statistic used in this analysis is the lasso coefficient difference statistic.
}
\references{
Srinivasan, A, Xue, L, Zhan, X. Compositional knockoff filter for high-dimensional regression analysis of microbiome data.
Biometrics. 2020; 1– 12. https://doi.org/10.1111/biom.13336
}
