% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRKF.R
\name{multilayerKF}
\alias{multilayerKF}
\title{Run the multiple response knockoff filter}
\usage{
multilayerKF(W, Y, q, penalty = "lasso")
}
\arguments{
\item{W}{OTU count matrix}

\item{Y}{matrix of responses}

\item{q}{target FDP}

\item{penalty}{type of penalty for selection}
}
\description{
Run the multiple response knockoff filter
}
