% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multRegression.R
\name{SCADSelection}
\alias{SCADSelection}
\title{Run the selection method through SCAD penalty (Fan, 2001)}
\usage{
SCADSelection(Xaug, Y, n, Omega0 = NULL)
}
\arguments{
\item{Xaug}{Design matrix of original data column augmented by knockoffs}

\item{Y}{vectorized response vector}

\item{n}{the number of samples}

\item{Omega0}{the estimated omega matrix}
}
\value{
The result from the SCAD fit
}
\description{
Run the selection method through SCAD penalty (Fan, 2001)
}
