% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multRegression.R
\name{Step1}
\alias{Step1}
\title{Run Step 1 of Multivariate Outcome Regression method proposed by Sofer et al (2014)}
\usage{
Step1(Xaug, Y, n, K, method = "S")
}
\arguments{
\item{Xaug}{Design matrix of original data column augmented by knockoffs}

\item{Y}{vectorized response vector}

\item{n}{number of samples}

\item{K}{number of responses}
}
\value{
the omega estimated through step 1
}
\description{
Run Step 1 of Multivariate Outcome Regression method proposed by Sofer et al (2014)
}
