% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multRegression.R
\name{Step2}
\alias{Step2}
\title{Run Step 2 of Multivariate Outcome Regression method proposed by Sofer et al (2014)}
\usage{
Step2(Xaug, Y, step1Omega, n, penalty = "scad")
}
\arguments{
\item{Xaug}{Design matrix of original data column augmented by knockoffs}

\item{Y}{vectorized response vector}

\item{step1Omega}{the omega estimated through step 1}

\item{n}{the number of samples}

\item{penalty}{the penalty function of interest}
}
\value{
The result after selection
}
\description{
Run Step 2 of Multivariate Outcome Regression method proposed by Sofer et al (2014)
}
